Inline es_serde_1 serialiser
